Gateway-App/src/test/javascript/spec/app/entities/SmartLernplatz/learning-spot/learning-spot.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { LearningSpotService } from 'app/entities/SmartLernplatz/learning-spot/learning-spot.service';
import { ILearningSpot, LearningSpot } from 'app/shared/model/SmartLernplatz/learning-spot.model';

describe('Service Tests', () => {
  describe('LearningSpot Service', () => {
    let injector: TestBed;
    let service: LearningSpotService;
    let httpMock: HttpTestingController;
    let elemDefault: ILearningSpot;
    let expectedResult: ILearningSpot | ILearningSpot[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(LearningSpotService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new LearningSpot(0, false, false, false);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a LearningSpot', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new LearningSpot()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a LearningSpot', () => {
        const returnedFromService = Object.assign(
          {
            occupied: true,
            socket: true,
            silence: true
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of LearningSpot', () => {
        const returnedFromService = Object.assign(
          {
            occupied: true,
            socket: true,
            silence: true
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a LearningSpot', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
